refactor logic and miscellaneous fixes
