+++ public/components/datasources/components/manage/associated_objects/associated_objects_tab.tsx
@@ -6,41 +6,29 @@
 import React, { useEffect, useState } from 'react';
 import _ from 'lodash';
 import {
-  EuiInMemoryTable,
-  EuiLink,
   EuiPanel,
   EuiFlexGroup,
   EuiFlexItem,
   EuiText,
   EuiHorizontalRule,
   EuiSpacer,
-  SearchFilterConfig,
-  EuiTableFieldDataColumnType,
   EuiSelectable,
 } from '@elastic/eui';
 import { i18n } from '@osd/i18n';
 import {
-  AccelerationIndexType,
   AssociatedObject,
   CachedAcceleration,
+  CachedAccelerationByDataSource,
+  CachedDataSourceStatus,
   CachedDatabase,
   CachedTable,
 } from '../../../../../../common/types/data_connections';
-import {
-  getRenderAccelerationDetailsFlyout,
-  getRenderAssociatedObjectsDetailsFlyout,
-} from '../../../../../plugin';
 import { AccelerationsRecommendationCallout } from './accelerations_recommendation_callout';
 import {
-  ASSC_OBJ_TABLE_ACC_COLUMN_NAME,
-  ASSC_OBJ_TABLE_SEARCH_HINT,
   ASSC_OBJ_PANEL_TITLE,
-  ASSC_OBJ_PANEL_DESRIPTION,
+  ASSC_OBJ_PANEL_DESCRIPTION,
   ASSC_OBJ_FRESH_MSG,
-  ASSC_OBJ_TABLE_SUBJ,
-  isTablesCacheUpdated,
-  isAccelerationsCacheUpdated,
-  isDatabasesCacheUpdated,
+  getAccelerationType,
 } from './utils/associated_objects_tab_utils';
 import { DatasourceDetails } from '../data_connection';
 import { DirectQueryLoadingStatus } from '../../../../../../common/types/explorer';
@@ -48,55 +36,38 @@ import { AssociatedObjectsTabEmpty } from './utils/associated_objects_tab_empty'
 import { AssociatedObjectsTabLoading } from './utils/associated_objects_tab_loading';
 import { AssociatedObjectsRefreshButton } from './utils/associated_objects_refresh_button';
 import { CatalogCacheManager } from '../../../../../../public/framework/catalog_cache/cache_manager';
+import { AssociatedObjectsTable } from './modules/associated_objects_table';
+import {
+  useLoadAccelerationsToCache,
+  useLoadDatabasesToCache,
+  useLoadTablesToCache,
+} from '../../../../../../public/framework/catalog_cache/cache_loader';
 
 export interface AssociatedObjectsTabProps {
   datasource: DatasourceDetails;
-  databasesLoadStatus: DirectQueryLoadingStatus;
-  startLoadingDatabases: (datasource: string) => void;
-  cachedDatabases: CachedDatabase[];
-  selectedDatabase: string;
-  setSelectedDatabase: React.Dispatch<React.SetStateAction<string>>;
-  tablesLoadStatus: DirectQueryLoadingStatus;
-  startLoadingTables: (datasource: string, database?: string) => void;
-  accelerationsLoadStatus: DirectQueryLoadingStatus;
-  startLoadingAccelerations: (datasource: string) => void;
-  isFirstTimeLoading: boolean;
-  isRefreshing: boolean;
-  setIsRefreshing: React.Dispatch<React.SetStateAction<boolean>>;
-}
-
-interface FilterOption {
-  value: string;
-  text: string;
-}
-
-interface AssociatedTableFilter {
-  type: string;
-  field: string;
-  operator: string;
-  value: string;
 }
 
 export const AssociatedObjectsTab: React.FC<AssociatedObjectsTabProps> = (props) => {
-  const {
-    datasource,
-    databasesLoadStatus,
-    startLoadingDatabases,
-    cachedDatabases,
-    selectedDatabase,
-    setSelectedDatabase,
-    tablesLoadStatus,
-    startLoadingTables,
-    accelerationsLoadStatus,
-    startLoadingAccelerations,
-    isFirstTimeLoading,
-    isRefreshing,
-    setIsRefreshing,
-  } = props;
+  const { datasource } = props;
+  const [selectedDatabase, setSelectedDatabase] = useState<string>('');
+  const [isRefreshing, setIsRefreshing] = useState<boolean>(false);
   const [lastUpdated, setLastUpdated] = useState(new Date().toLocaleString());
   const [isObjectsLoading, setIsObjectsLoading] = useState<boolean>(false);
+  const [cachedDatabases, setCachedDatabases] = useState<CachedDatabase[]>([]);
   const [cachedTables, setCachedTables] = useState<CachedTable[]>([]);
   const [cachedAccelerations, setCachedAccelerations] = useState<CachedAcceleration[]>([]);
+  const [associatedObjects, setAssociatedObjects] = useState<AssociatedObject[]>([]);
+  const [isFirstTimeLoading, setIsFirstTimeLoading] = useState<boolean>(true);
+
+  const {
+    loadStatus: databasesLoadStatus,
+    startLoading: startLoadingDatabases,
+  } = useLoadDatabasesToCache();
+  const { loadStatus: tablesLoadStatus, startLoading: startLoadingTables } = useLoadTablesToCache();
+  const {
+    loadStatus: accelerationsLoadStatus,
+    startLoading: startLoadingAccelerations,
+  } = useLoadAccelerationsToCache();
 
   let lastChecked: boolean;
   if (selectedDatabase !== '') {
@@ -121,10 +92,9 @@ export const AssociatedObjectsTab: React.FC<AssociatedObjectsTabProps> = (props)
   );
 
   const onRefreshButtonClick = () => {
-    console.log('clicked on refresh button, i will update implementation later');
-    CatalogCacheManager.clearAccelerationsCache();
-    CatalogCacheManager.clearDataSourceCache();
     startLoadingDatabases(datasource.name);
+    startLoadingTables(datasource.name, selectedDatabase);
+    startLoadingAccelerations(datasource.name);
     setIsRefreshing(true);
   };
 
@@ -134,7 +104,7 @@ export const AssociatedObjectsTab: React.FC<AssociatedObjectsTabProps> = (props)
     });
 
     const panelDescription = i18n.translate('datasources.associatedObjectsTab.panelDescription', {
-      defaultMessage: ASSC_OBJ_PANEL_DESRIPTION,
+      defaultMessage: ASSC_OBJ_PANEL_DESCRIPTION,
     });
 
     const LastUpdatedText = () => {
@@ -168,326 +138,83 @@ export const AssociatedObjectsTab: React.FC<AssociatedObjectsTabProps> = (props)
     );
   };
 
-  const AssociatedObjectsTable = () => {
-    const [databaseFilterOptions, setDatabaseFilterOptions] = useState<FilterOption[]>([]);
-    const [accelerationFilterOptions, setAccelerationFilterOptions] = useState<FilterOption[]>([]);
-    const [filteredObjects, setFilteredObjects] = useState<AssociatedObject[]>([]);
-    const [associatedObjects, setAssociatedObjects] = useState<AssociatedObject[]>([]);
-
-    const columns = [
-      {
-        field: 'name',
-        name: i18n.translate('datasources.associatedObjectsTab.column.name', {
-          defaultMessage: 'Name',
-        }),
-        sortable: true,
-        'data-test-subj': 'nameCell',
-        render: (name: string, item: AssociatedObject) => (
-          <EuiLink
-            onClick={() => {
-              if (item.type === 'Table') {
-                renderAssociatedObjectsDetailsFlyout(item);
-              } else {
-                renderAccelerationDetailsFlyout(item.accelerations[0]);
-              }
-            }}
-          >
-            {name}
-          </EuiLink>
-        ),
-      },
-      {
-        field: 'database',
-        name: i18n.translate('datasources.associatedObjectsTab.column.database', {
-          defaultMessage: 'Database',
-        }),
-        sortable: true,
-      },
-      {
-        field: 'type',
-        name: i18n.translate('datasources.associatedObjectsTab.column.type', {
-          defaultMessage: 'Type',
-        }),
-        sortable: true,
-      },
-      {
-        field: 'createdByIntegration',
-        name: i18n.translate('datasources.associatedObjectsTab.column.createdByIntegration', {
-          defaultMessage: 'Created by Integration',
-        }),
-        sortable: true,
-        render: (createdByIntegration: string, _item: AssociatedObject) =>
-          createdByIntegration ? (
-            <EuiLink onClick={() => openDetailsPage(createdByIntegration)}>
-              {createdByIntegration}
-            </EuiLink>
-          ) : (
-            '-'
-          ),
-      },
-      {
-        field: 'accelerations',
-        name: i18n.translate('datasources.associatedObjectsTab.column.accelerations', {
-          defaultMessage: 'Accelerations',
-        }),
-        sortable: true,
-        render: (accelerations: string[]) => {
-          return accelerations.length > 0
-            ? accelerations.map((acceleration, index) => (
-                <React.Fragment key={index}>
-                  <EuiLink onClick={() => renderAccelerationDetailsFlyout(acceleration)}>
-                    {acceleration.name}
-                  </EuiLink>
-                  {index < accelerations.length - 1 ? ', ' : ''}
-                </React.Fragment>
-              ))
-            : '-';
-        },
-      },
-      {
-        name: i18n.translate('datasources.associatedObjectsTab.column.actions', {
-          defaultMessage: 'Actions',
-        }),
-        actions: [
-          {
-            name: i18n.translate('datasources.associatedObjectsTab.action.discover.name', {
-              defaultMessage: 'Discover',
-            }),
-            description: i18n.translate(
-              'datasources.associatedObjectsTab.action.discover.description',
-              {
-                defaultMessage: 'Discover this object',
-              }
-            ),
-            type: 'icon',
-            icon: 'discoverApp',
-            onClick: (item: AssociatedObject) => console.log('Discover', item),
-          },
-          {
-            name: i18n.translate('datasources.associatedObjectsTab.action.accelerate.name', {
-              defaultMessage: 'Accelerate',
-            }),
-            description: i18n.translate(
-              'datasources.associatedObjectsTab.action.accelerate.description',
-              {
-                defaultMessage: 'Accelerate this object',
-              }
-            ),
-            type: 'icon',
-            icon: 'bolt',
-            available: (item: AssociatedObject) => item.type === 'Table',
-            onClick: (item: AssociatedObject) => console.log('Accelerate', item),
-          },
-        ],
-      },
-    ] as Array<EuiTableFieldDataColumnType<any>>;
-
-    const onSearchChange = ({ query, error }) => {
-      if (error) {
-        console.log('Search error:', error);
-        return;
-      }
-
-      const matchesClauses = (obj: AssociatedObject, clauses: AssociatedTableFilter[]): boolean => {
-        if (clauses.length === 0) return true;
-
-        return clauses.some((clause) => {
-          if (clause.field !== ASSC_OBJ_TABLE_ACC_COLUMN_NAME) {
-            return obj[clause.field] === clause.value;
-          } else if (
-            clause.field === ASSC_OBJ_TABLE_ACC_COLUMN_NAME &&
-            Array.isArray(obj.accelerations)
-          ) {
-            return obj.accelerations.some((acceleration) => acceleration.name === clause.value);
-          }
-
-          return false;
-        });
-      };
-
-      const filtered = associatedObjects.filter((obj) => {
-        const clauses = query.ast._clauses;
-        return matchesClauses(obj, clauses);
-      });
-
-      setFilteredObjects(filtered);
-    };
-
-    const searchFilters = [
-      {
-        type: 'field_value_selection',
-        field: 'database',
-        name: 'Database',
-        multiSelect: true,
-        options: databaseFilterOptions,
-        cache: 60000,
-      },
-      {
-        type: 'field_value_selection',
-        field: 'accelerations',
-        name: 'Accelerations',
-        multiSelect: true,
-        options: accelerationFilterOptions,
-        cache: 60000,
-      },
-    ] as SearchFilterConfig[];
-
-    const tableSearch = {
-      filters: searchFilters,
-      box: {
-        incremental: true,
-        placeholder: ASSC_OBJ_TABLE_SEARCH_HINT,
-        schema: {
-          fields: { name: { type: 'string' }, database: { type: 'string' } },
-        },
-      },
-      onChange: onSearchChange,
-    };
-
-    const pagination = {
-      initialPageSize: 10,
-      pageSizeOptions: [10, 25, 50],
-    };
-
-    const sorting = {
-      sort: {
-        field: 'name',
-        direction: 'asc',
-      },
-    };
-
-    const getAccelerationType = (type: AccelerationIndexType) => {
-      switch (type) {
-        case 'skipping':
-          return 'Skipping Index';
-        case 'covering':
-          return 'Covering Index';
-        case 'materialized':
-          return 'Materialized View';
-      }
-    };
-
-    useEffect(() => {
-      console.log('component mounted');
-      console.log(
-        'cached accelerations',
-        cachedAccelerations.filter(
-          (acceleration: CachedAcceleration) => acceleration.database === selectedDatabase
-        )
-      );
-      const tableObjects: AssociatedObject[] = cachedTables.map((table: CachedTable) => {
-        return {
-          datasource: datasource.name,
-          id: table.name,
-          name: table.name,
-          database: selectedDatabase,
-          type: 'table',
-          createdByIntegration: 'N/A',
-          // Temporary dummy array
-          accelerations: [],
-          columns: table.columns,
-        };
-      });
-      const accelerationObjects: AssociatedObject[] = cachedAccelerations
-        .filter((acceleration: CachedAcceleration) => acceleration.database === selectedDatabase)
-        .map((acceleration: CachedAcceleration) => ({
-          datasource: datasource.name,
-          id: acceleration.flintIndexName,
-          name: acceleration.flintIndexName,
-          database: selectedDatabase,
-          type: getAccelerationType(acceleration.type),
-          createdByIntegration: '-',
-          // Temporary dummy array
-          accelerations: [],
-          columns: undefined,
-        }));
-      console.log(accelerationObjects);
-      setAssociatedObjects([...tableObjects, ...accelerationObjects]);
-    }, []);
-
-    useEffect(() => {
-      const databaseOptions = Array.from(new Set(associatedObjects.map((obj) => obj.database)))
-        .sort()
-        .map((database) => ({ value: database, text: database }));
-      setDatabaseFilterOptions(databaseOptions);
-
-      const accelerationOptions = Array.from(
-        new Set(
-          associatedObjects
-            .flatMap((obj) => obj.accelerations.map((acceleration) => acceleration.name))
-            .filter(Boolean)
-        )
-      )
-        .sort()
-        .map((name) => ({ value: name, text: name }));
-      setAccelerationFilterOptions(accelerationOptions);
-      setFilteredObjects(associatedObjects);
-
-      return () => {
-        console.log('component unmounted');
-      };
-    }, [associatedObjects]);
-
-    return (
-      <EuiInMemoryTable
-        items={filteredObjects}
-        columns={columns}
-        search={tableSearch}
-        pagination={pagination}
-        sorting={sorting}
-        data-test-subj={ASSC_OBJ_TABLE_SUBJ}
-      />
-    );
-  };
-
+  // Load databases if empty or retrieve from cache if updated
   useEffect(() => {
-    // Reload tables and accelerations to cache if nothing in cache
-    console.log('in here');
-    if (selectedDatabase && isDatabasesCacheUpdated(datasource.name)) {
-      const tablesCache = isTablesCacheUpdated(datasource.name, selectedDatabase);
-      const accelerationsCache = isAccelerationsCacheUpdated();
-      console.log(
-        'tablesCache',
-        tablesCache,
-        'accelerationsCache',
-        accelerationsCache,
-        'isRefreshing',
-        isRefreshing
-      );
-      if (!tablesCache || !accelerationsCache) {
-        console.log('either tables or accelerations cache not updated');
-        if (!tablesCache) {
-          startLoadingTables(datasource.name, selectedDatabase);
-        }
-        if (!accelerationsCache) startLoadingAccelerations(datasource.name);
-        setIsObjectsLoading(true);
-      } else if (tablesCache && accelerationsCache) {
-        console.log('tables and acceleration cache are updated');
-        setCachedTables(CatalogCacheManager.getDatabase(datasource.name, selectedDatabase).tables);
-        setCachedAccelerations(CatalogCacheManager.getAccelerationsCache().accelerations);
-        setLastUpdated(
-          CatalogCacheManager.getDatabase(datasource.name, selectedDatabase).lastUpdated
-        );
-        if (
-          tablesLoadStatus === DirectQueryLoadingStatus.SUCCESS &&
-          accelerationsLoadStatus === DirectQueryLoadingStatus.SUCCESS
-        ) {
-          setIsRefreshing(false);
-        }
-        setIsObjectsLoading(false);
+    if (datasource.name) {
+      const datasourceCache = CatalogCacheManager.getOrCreateDataSource(datasource.name);
+      if (datasourceCache.status === CachedDataSourceStatus.Empty) {
+        startLoadingDatabases(datasource.name);
+      } else if (datasourceCache.status === CachedDataSourceStatus.Updated) {
+        setCachedDatabases(datasourceCache.databases);
+        setIsFirstTimeLoading(false);
       }
     }
-  }, [
-    isRefreshing,
-    selectedDatabase,
-    databasesLoadStatus,
-    tablesLoadStatus,
-    accelerationsLoadStatus,
-  ]);
+  }, [datasource.name]);
 
+  // Retrieve from cache upon load success
   useEffect(() => {
-    setSelectedDatabase(databaseSelectorOptions.find((option) => option.checked === 'on')?.label);
-  }, [databaseSelectorOptions]);
+    const status = databasesLoadStatus.toLowerCase();
+    const datasourceCache = CatalogCacheManager.getOrCreateDataSource(datasource.name);
+    if (status === DirectQueryLoadingStatus.SUCCESS) {
+      setCachedDatabases(datasourceCache.databases);
+      setIsFirstTimeLoading(false);
+    }
+  }, [datasource.name, databasesLoadStatus]);
+
+  const handleObjectsLoad = (
+    databaseCache: CachedDatabase,
+    accelerationsCache: CachedAccelerationByDataSource
+  ) => {
+    if (
+      databaseCache.status === CachedDataSourceStatus.Updated &&
+      accelerationsCache.status === CachedDataSourceStatus.Updated
+    ) {
+      setLastUpdated(new Date(databaseCache.lastUpdated).toLocaleString());
+      setIsRefreshing(false);
+      setIsObjectsLoading(false);
+    }
+  };
+
+  // Load tables and accelerations if empty or retrieve from cache if not
+  useEffect(() => {
+    if (datasource.name && selectedDatabase) {
+      const databaseCache = CatalogCacheManager.getDatabase(datasource.name, selectedDatabase);
+      const accelerationsCache = CatalogCacheManager.getOrCreateAccelerationsByDataSource(
+        datasource.name
+      );
+      if (databaseCache.status === CachedDataSourceStatus.Empty) {
+        startLoadingTables(datasource.name, selectedDatabase);
+        setIsObjectsLoading(true);
+      } else if (databaseCache.status === CachedDataSourceStatus.Updated) {
+        setCachedTables(databaseCache.tables);
+      }
+      if (accelerationsCache.status === CachedDataSourceStatus.Empty) {
+        startLoadingAccelerations(datasource.name);
+        setIsObjectsLoading(true);
+      } else if (accelerationsCache.status === CachedDataSourceStatus.Updated) {
+        setCachedAccelerations(accelerationsCache.accelerations);
+      }
+    }
+  }, [datasource.name, selectedDatabase]);
+
+  // Retrieve from tables cache upon load success
+  useEffect(() => {
+    if (datasource.name && selectedDatabase) {
+      const tablesStatus = tablesLoadStatus.toLowerCase();
+      const databaseCache = CatalogCacheManager.getDatabase(datasource.name, selectedDatabase);
+      const accelerationsStatus = accelerationsLoadStatus.toLowerCase();
+      const accelerationsCache = CatalogCacheManager.getOrCreateAccelerationsByDataSource(
+        datasource.name
+      );
+      if (tablesStatus === DirectQueryLoadingStatus.SUCCESS) {
+        setCachedTables(databaseCache.tables);
+      }
+      if (accelerationsStatus === DirectQueryLoadingStatus.SUCCESS) {
+        setCachedAccelerations(accelerationsCache.accelerations);
+      }
+      handleObjectsLoad(databaseCache, accelerationsCache);
+    }
+  }, [datasource.name, selectedDatabase, tablesLoadStatus, accelerationsLoadStatus]);
 
   useEffect(() => {
     setDatabaseSelectorOptions(
@@ -507,11 +234,38 @@ export const AssociatedObjectsTab: React.FC<AssociatedObjectsTabProps> = (props)
   }, [cachedDatabases]);
 
   useEffect(() => {
-    console.log('selectedDatabase', selectedDatabase);
-  }, [selectedDatabase]);
+    setSelectedDatabase(databaseSelectorOptions.find((option) => option.checked === 'on')?.label);
+  }, [databaseSelectorOptions]);
 
-  const renderAccelerationDetailsFlyout = getRenderAccelerationDetailsFlyout();
-  const renderAssociatedObjectsDetailsFlyout = getRenderAssociatedObjectsDetailsFlyout();
+  useEffect(() => {
+    const tableObjects: AssociatedObject[] = cachedTables.map((table: CachedTable) => {
+      return {
+        datasource: datasource.name,
+        id: table.name,
+        name: table.name,
+        database: selectedDatabase,
+        type: 'table',
+        createdByIntegration: 'N/A',
+        // Temporary dummy array
+        accelerations: [],
+        columns: table.columns,
+      };
+    });
+    const accelerationObjects: AssociatedObject[] = cachedAccelerations
+      .filter((acceleration: CachedAcceleration) => acceleration.database === selectedDatabase)
+      .map((acceleration: CachedAcceleration) => ({
+        datasource: datasource.name,
+        id: acceleration.flintIndexName,
+        name: acceleration.flintIndexName,
+        database: acceleration.database,
+        type: getAccelerationType(acceleration.type),
+        createdByIntegration: '-',
+        // Temporary dummy array
+        accelerations: [],
+        columns: undefined,
+      }));
+    setAssociatedObjects([...tableObjects, ...accelerationObjects]);
+  }, [selectedDatabase, cachedTables, cachedAccelerations]);
 
   return (
     <>
@@ -523,7 +277,7 @@ export const AssociatedObjectsTab: React.FC<AssociatedObjectsTabProps> = (props)
           <AssociatedObjectsTabLoading objectType="databases" warningMessage={false} />
         ) : (
           <>
-            {isDatabasesCacheUpdated(datasource.name) && cachedDatabases.length === 0 ? (
+            {cachedDatabases.length === 0 ? (
               <AssociatedObjectsTabEmpty cacheType="databases" />
             ) : (
               <>
@@ -547,14 +301,18 @@ export const AssociatedObjectsTab: React.FC<AssociatedObjectsTabProps> = (props)
                     </EuiSelectable>
                   </EuiFlexItem>
                   <EuiFlexItem>
-                    {isObjectsLoading ? (
+                    {isObjectsLoading && !isRefreshing ? (
                       <AssociatedObjectsTabLoading objectType="tables" warningMessage={true} />
                     ) : (
                       <>
-                        {cachedTables.length === 0 || cachedAccelerations.length === 0 ? (
-                          <AssociatedObjectsTabEmpty cacheType="tables" />
+                        {cachedTables.length > 0 || cachedAccelerations.length > 0 ? (
+                          <AssociatedObjectsTable
+                            datasource={datasource}
+                            selectedDatabase={selectedDatabase}
+                            associatedObjects={associatedObjects}
+                          />
                         ) : (
-                          <AssociatedObjectsTable />
+                          <AssociatedObjectsTabEmpty cacheType="tables" />
                         )}
                       </>
                     )}
