+++ public/components/datasources/components/manage/associated_objects/modules/associated_objects_table.tsx
@@ -1,0 +1,253 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import React, { useEffect, useState } from 'react';
+import _ from 'lodash';
+import {
+  EuiInMemoryTable,
+  EuiLink,
+  SearchFilterConfig,
+  EuiTableFieldDataColumnType,
+} from '@elastic/eui';
+import { i18n } from '@osd/i18n';
+import { AssociatedObject } from '../../../../../../../common/types/data_connections';
+import {
+  getRenderAccelerationDetailsFlyout,
+  getRenderAssociatedObjectsDetailsFlyout,
+} from '../../../../../../plugin';
+import {
+  ASSC_OBJ_TABLE_ACC_COLUMN_NAME,
+  ASSC_OBJ_TABLE_SEARCH_HINT,
+  ASSC_OBJ_TABLE_SUBJ,
+} from '../utils/associated_objects_tab_utils';
+import { DatasourceDetails } from '../../data_connection';
+
+interface AssociatedObjectsTableProps {
+  datasource: DatasourceDetails;
+  selectedDatabase: string;
+  associatedObjects: AssociatedObject[];
+}
+
+interface FilterOption {
+  value: string;
+  text: string;
+}
+
+interface AssociatedTableFilter {
+  type: string;
+  field: string;
+  operator: string;
+  value: string;
+}
+
+export const AssociatedObjectsTable = (props: AssociatedObjectsTableProps) => {
+  const { associatedObjects } = props;
+  const [accelerationFilterOptions, setAccelerationFilterOptions] = useState<FilterOption[]>([]);
+  const [filteredObjects, setFilteredObjects] = useState<AssociatedObject[]>([]);
+
+  const columns = [
+    {
+      field: 'name',
+      name: i18n.translate('datasources.associatedObjectsTab.column.name', {
+        defaultMessage: 'Name',
+      }),
+      sortable: true,
+      'data-test-subj': 'nameCell',
+      render: (name: string, item: AssociatedObject) => (
+        <EuiLink
+          onClick={() => {
+            console.log('name', name);
+            console.log('item', item);
+            if (item.type === 'table') {
+              renderAssociatedObjectsDetailsFlyout(item);
+            } else {
+              renderAccelerationDetailsFlyout(item);
+            }
+          }}
+        >
+          {name}
+        </EuiLink>
+      ),
+    },
+    {
+      field: 'database',
+      name: i18n.translate('datasources.associatedObjectsTab.column.database', {
+        defaultMessage: 'Database',
+      }),
+      sortable: true,
+    },
+    {
+      field: 'type',
+      name: i18n.translate('datasources.associatedObjectsTab.column.type', {
+        defaultMessage: 'Type',
+      }),
+      sortable: true,
+    },
+    {
+      field: 'createdByIntegration',
+      name: i18n.translate('datasources.associatedObjectsTab.column.createdByIntegration', {
+        defaultMessage: 'Created by Integration',
+      }),
+      sortable: true,
+      render: (createdByIntegration: string, _item: AssociatedObject) =>
+        createdByIntegration ? (
+          <EuiLink onClick={() => openDetailsPage(createdByIntegration)}>
+            {createdByIntegration}
+          </EuiLink>
+        ) : (
+          '-'
+        ),
+    },
+    {
+      field: 'accelerations',
+      name: i18n.translate('datasources.associatedObjectsTab.column.accelerations', {
+        defaultMessage: 'Accelerations',
+      }),
+      sortable: true,
+      render: (accelerations: string[]) => {
+        return accelerations.length > 0
+          ? accelerations.map((acceleration, index) => (
+              <React.Fragment key={index}>
+                <EuiLink onClick={() => renderAccelerationDetailsFlyout(acceleration)}>
+                  {acceleration.name}
+                </EuiLink>
+                {index < accelerations.length - 1 ? ', ' : ''}
+              </React.Fragment>
+            ))
+          : '-';
+      },
+    },
+    {
+      name: i18n.translate('datasources.associatedObjectsTab.column.actions', {
+        defaultMessage: 'Actions',
+      }),
+      actions: [
+        {
+          name: i18n.translate('datasources.associatedObjectsTab.action.discover.name', {
+            defaultMessage: 'Discover',
+          }),
+          description: i18n.translate(
+            'datasources.associatedObjectsTab.action.discover.description',
+            {
+              defaultMessage: 'Discover this object',
+            }
+          ),
+          type: 'icon',
+          icon: 'discoverApp',
+          onClick: (item: AssociatedObject) => console.log('Discover', item),
+        },
+        {
+          name: i18n.translate('datasources.associatedObjectsTab.action.accelerate.name', {
+            defaultMessage: 'Accelerate',
+          }),
+          description: i18n.translate(
+            'datasources.associatedObjectsTab.action.accelerate.description',
+            {
+              defaultMessage: 'Accelerate this object',
+            }
+          ),
+          type: 'icon',
+          icon: 'bolt',
+          available: (item: AssociatedObject) => item.type === 'Table',
+          onClick: (item: AssociatedObject) => console.log('Accelerate', item),
+        },
+      ],
+    },
+  ] as Array<EuiTableFieldDataColumnType<any>>;
+
+  const onSearchChange = ({ query, error }) => {
+    if (error) {
+      console.log('Search error:', error);
+      return;
+    }
+
+    const matchesClauses = (obj: AssociatedObject, clauses: AssociatedTableFilter[]): boolean => {
+      if (clauses.length === 0) return true;
+
+      return clauses.some((clause) => {
+        if (clause.field !== ASSC_OBJ_TABLE_ACC_COLUMN_NAME) {
+          return obj[clause.field] === clause.value;
+        } else if (
+          clause.field === ASSC_OBJ_TABLE_ACC_COLUMN_NAME &&
+          Array.isArray(obj.accelerations)
+        ) {
+          return obj.accelerations.some((acceleration) => acceleration.name === clause.value);
+        }
+
+        return false;
+      });
+    };
+
+    const filtered = associatedObjects.filter((obj) => {
+      const clauses = query.ast._clauses;
+      return matchesClauses(obj, clauses);
+    });
+
+    setFilteredObjects(filtered);
+  };
+
+  const searchFilters = [
+    {
+      type: 'field_value_selection',
+      field: 'accelerations',
+      name: 'Accelerations',
+      multiSelect: true,
+      options: accelerationFilterOptions,
+      cache: 60000,
+    },
+  ] as SearchFilterConfig[];
+
+  const tableSearch = {
+    filters: searchFilters,
+    box: {
+      incremental: true,
+      placeholder: ASSC_OBJ_TABLE_SEARCH_HINT,
+      schema: {
+        fields: { name: { type: 'string' }, database: { type: 'string' } },
+      },
+    },
+    onChange: onSearchChange,
+  };
+
+  const pagination = {
+    initialPageSize: 10,
+    pageSizeOptions: [10, 25, 50],
+  };
+
+  const sorting = {
+    sort: {
+      field: 'name',
+      direction: 'asc',
+    },
+  };
+
+  useEffect(() => {
+    const accelerationOptions = Array.from(
+      new Set(
+        associatedObjects
+          .flatMap((obj) => obj.accelerations.map((acceleration) => acceleration.name))
+          .filter(Boolean)
+      )
+    )
+      .sort()
+      .map((name) => ({ value: name, text: name }));
+    setAccelerationFilterOptions(accelerationOptions);
+    setFilteredObjects(associatedObjects);
+  }, [associatedObjects]);
+
+  const renderAccelerationDetailsFlyout = getRenderAccelerationDetailsFlyout();
+  const renderAssociatedObjectsDetailsFlyout = getRenderAssociatedObjectsDetailsFlyout();
+
+  return (
+    <EuiInMemoryTable
+      items={filteredObjects}
+      columns={columns}
+      search={tableSearch}
+      pagination={pagination}
+      sorting={sorting}
+      data-test-subj={ASSC_OBJ_TABLE_SUBJ}
+    />
+  );
+};
